tl: pass the calculated digest to the server
As a way of protecting the service from DDoS or other abuse the clients
will send their expected digest in a query parameter. If this is a new
entry this query parameter can be used to measure if a client is
requesting new entries without knowing the correct digest ahead of time.