README: update 'tl cat' example to work again
The mutable 'master' reference has mutated since it was added to the
readme, so unsurprisingly the example prints an error now.

Update the example to something that is expected to verify correctly,
and hopefully provide a better example of usage.

The example project chosen doesn't tag releases, so we're stuck with an
opaque commit hash.